delete message confirm cancel

+++ src/index.js
@@ -1,44 +1,45 @@
 import "./assets/styles/styles.scss";
 import "./index.scss";
+import { openModal } from "./assets/javascripts/modal";
 
 const articleContainerElement = document.querySelector(".articles-container");
 const categoriesContainerElement = document.querySelector(".categories");
-const selectElements = document.querySelector('select')
+const selectElement = document.querySelector("select");
 let filter;
 let articles;
 let sortBy = "desc";
 
-selectElements.addEventListener("change", ()=>{
-    sortBy = selectElements.value
-    fetchArticle()
-})
+selectElement.addEventListener("change", () => {
+  sortBy = selectElement.value;
+  fetchArticle();
+});
 
 const createArticles = () => {
   const articlesDOM = articles
-    .filter(article => {
-      if (filter) {
-        return article.category === filter;
-      } else {
-        return true;
-      }
-    })
-    .map(article => {
-      const articleDOM = document.createElement("div");
-      articleDOM.classList.add("article");
-      articleDOM.innerHTML = `
+      .filter(article => {
+        if (filter) {
+          return article.category === filter;
+        } else {
+          return true;
+        }
+      })
+      .map(article => {
+        const articleDOM = document.createElement("div");
+        articleDOM.classList.add("article");
+        articleDOM.innerHTML = `
 <img
   src="${article.img}"
   alt="profile"
 />
 <h2>${article.title}</h2>
 <p class="article-author">${article.author} - ${new Date(
-        article.createdAt
-      ).toLocaleDateString("fr-FR", {
-        weekday: "long",
-        day: "2-digit",
-        month: "long",
-        year: "numeric"
-      })}</p>
+            article.createdAt
+        ).toLocaleDateString("fr-FR", {
+          weekday: "long",
+          day: "2-digit",
+          month: "long",
+          year: "numeric"
+        })}</p>
 <p class="article-content">
   ${article.content}
 </p>
@@ -47,8 +48,8 @@ const createArticles = () => {
   <button class="btn btn-primary" data-id=${article._id} >Modifier</button>
 </div>
 `;
-      return articleDOM;
-    });
+        return articleDOM;
+      });
   articleContainerElement.innerHTML = "";
   articleContainerElement.append(...articlesDOM);
   const deleteButtons = articleContainerElement.querySelectorAll(".btn-danger");
@@ -57,25 +58,29 @@ const createArticles = () => {
     button.addEventListener("click", event => {
       const target = event.target;
       const articleId = target.dataset.id;
-      window.location.assign(`./form.html?id=${articleId}`);
+      location.assign(`/form.html?id=${articleId}`);
     });
   });
   deleteButtons.forEach(button => {
     button.addEventListener("click", async event => {
-      try {
-        const target = event.target;
-        const articleId = target.dataset.id;
-        const response = await fetch(
-          `https://restapi.fr/api/article/${articleId}`,
-          {
-            method: "DELETE"
-          }
-        );
-        const body = await response.json();
-        console.log(body);
-        fetchArticle();
-      } catch (e) {
-        console.log("e : ", e);
+      const result = await openModal(
+          "Etes vous sur de vouloir supprimer votre article ?"
+      );
+      if (result === true) {
+        try {
+          const target = event.target;
+          const articleId = target.dataset.id;
+          const response = await fetch(
+              `https://restapi.fr/api/article/${articleId}`,
+              {
+                method: "DELETE"
+              }
+          );
+          const body = await response.json();
+          fetchArticle();
+        } catch (e) {
+          console.log("e : ", e);
+        }
       }
     });
   });
@@ -85,21 +90,22 @@ const displayMenuCategories = categoriesArr => {
   const liElements = categoriesArr.map(categoryElem => {
     const li = document.createElement("li");
     li.innerHTML = `${categoryElem[0]} ( <strong>${categoryElem[1]}</strong> )`;
-    if(categoryElem[0] === filter){
-        li.classList.add('active');
+    if (categoryElem[0] === filter) {
+      li.classList.add("active");
     }
     li.addEventListener("click", () => {
       if (filter === categoryElem[0]) {
         filter = null;
         li.classList.remove("active");
+        createArticles();
       } else {
         filter = categoryElem[0];
         liElements.forEach(li => {
           li.classList.remove("active");
         });
         li.classList.add("active");
+        createArticles();
       }
-      createArticles();
     });
     return li;
   });
@@ -119,16 +125,18 @@ const createMenuCategories = () => {
   }, {});
 
   const categoriesArr = Object.keys(categories)
-    .map(category => {
-      return [category, categories[category]];
-    })
-    .sort((c1, c2) => c1[0].localeCompare(c2[0]));
+      .map(category => {
+        return [category, categories[category]];
+      })
+      .sort((c1, c2) => c1[0].localeCompare(c2[0]));
   displayMenuCategories(categoriesArr);
 };
 
 const fetchArticle = async () => {
   try {
-    const response = await fetch(`https://restapi.fr/api/article?sort=createAt:${sortBy}`);
+    const response = await fetch(
+        `https://restapi.fr/api/article?sort=createdAt:${sortBy}`
+    );
     articles = await response.json();
     createArticles();
     createMenuCategories();
